latest update
done

+++ app/src/main/java/com/arhamsoft/deskilz/networking/retrofit/URLConstant.kt
@@ -70,7 +70,7 @@ object URLConstant {
     var userImg: String = ""
     var u_id:String?="0"
     var chatHeadId:String?=""
-    var points:Int?=0
+    var points:Long?=0
     var player_count:Long? = 0L
     var joinType:Int = 9
     var currentLoginId:Int = 0
